+++ activerecord/CHANGELOG.md
@@ -1,1 +1,7 @@
+*   When calling the dynamic fixture accessor method with no arguments it now returns all fixtures of this type.
+    Previously this method always returned an empty array.
+
+    *Kevin McPhillips*
+
+
 Please check [5-1-stable](https://github.com/rails/rails/blob/5-1-stable/activerecord/CHANGELOG.md) for previous changes.